pyproject.toml
[tool.poetry]
name = "my-github-organizations"
version = "0.1.0"
description = ""
authors = ["insolor <insolor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.28.2"
pydantic = "^2.5.1"

[tool.poetry.group.dev.dependencies]
python-decouple = "^3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
run = "streamlit run app.py"
